+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- 🖼️ Added logoType (Asset, Network, None) support in QR Widget
+- 🖼️ Added default logo fallback when no logo is provided
+- 🛠️ Minor UI improvements in QR rendering
+
 ## 0.0.1
 
 - 🎉 Initial release